+++ includes/libraries/aihrus-framework/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## 1.2.3RC1
 * RESOLVE michael-cannon/testimonials-widget#184 Remove deprecated constructor calls for WP_Widget
-* Update Parsedown library
+* Update Parsedown 1.5.4 library
 
 ## 1.2.2
 * Add Inside Axelerant link